Updated Activity to Fragment and added Fragment code
