+++ app/src/main/java/com/mobewash/mobewash/PaymentFragment.java
@@ -7,7 +7,14 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.EditText;
 
+import com.stripe.android.Stripe;
+import com.stripe.android.TokenCallback;
+import com.stripe.android.model.Card;
+import com.stripe.android.model.Token;
+import com.stripe.android.view.CardInputWidget;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -18,14 +25,21 @@ import android.view.ViewGroup;
  * create an instance of this fragment.
  */
 public class PaymentFragment extends Fragment {
+
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
 
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
+    // Set up tag for logging
+    public static final String TAG = "Payment Fragment";
+
+    // Views
+    private EditText mCardInput;
+    private EditText mCardCVC;
+    private EditText mExpirationDate;
+    private Button mBookButton;
+    private EditText mBillingAddress;
 
     private OnFragmentInteractionListener mListener;
 
@@ -54,18 +68,37 @@ public class PaymentFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
+
+
+
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_payment, container, false);
-    }
+        View view =inflater.inflate(R.layout.fragment_payment, container, false);
+
+        // Set up Text Boxes
+        mCardInput = (EditText) getView().findViewById(R.id.CardNumber);
+        mCardCVC = (EditText) getView().findViewById(R.id.CVC);
+        mExpirationDate = (EditText) getView().findViewById(R.id.ExpirationDate);
+        mBillingAddress = (EditText) getView().findViewById(R.id.BillingAddress);
+
+        // Set up Book Button
+        mBookButton = (Button) getView().findViewById(R.id.BookButton);
+
+        //Set On Click Listener
+        mBookButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                String card= mCardInput.getText().toString();
+                String cvc= mCardCVC.getText().toString();
+                String exp= mExpirationDate.getText().toString();
+            }
+        });
+
+        return view;
+}
 
     // TODO: Rename method, update argument and hook method into UI event
     public void onButtonPressed(Uri uri) {
@@ -90,6 +123,41 @@ public class PaymentFragment extends Fragment {
         super.onDetach();
         mListener = null;
     }
+
+    /* Deciding which save card method to use*/
+    public void onClickSomething(String cardNumber, Integer cardExpMonth, Integer cardExpYear, String cardCVC) {
+        Card card = new Card(
+                cardNumber,
+                cardExpMonth,
+                cardExpYear,
+                cardCVC
+        );
+
+        card.validateNumber();
+        card.validateCVC();
+    }
+
+    /*
+    private void saveCard() {
+        Card card = mCardInputWidget.getCard();
+        if (card == null) {
+            return;
+        }
+
+        //Incorrect method call, needs a reference to context
+        Stripe stripe = new Stripe(getParent(), "pk_test_6pRNASCoBOKtIshFeQd4XMUh");
+        stripe.createToken(
+                card,
+                new TokenCallback() {
+                    public void onSuccess(Token token) {
+                        // Send token to your server
+                    }
+                    public void onError(Exception error) {
+                        // Show localized error message
+                    }
+                }
+        );
+    }*/
 
     /**
      * This interface must be implemented by activities that contain this
