+++ app/src/main/res/layout/activity_payment.xml
@@ -1,30 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:wallet="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:stripe="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.mobewash.mobewash.PaymentActivity">
 
     <fragment
-        android:id="@+id/wallet_fragment"
-        android:name="com.google.android.gms.wallet.fragment.SupportWalletFragment"
-        android:layout_width="368dp"
-        android:layout_height="wrap_content"
-        wallet:environment="test"
-        wallet:fragmentMode="buyButton"
-        tools:layout_editor_absoluteY="0dp"
-        tools:layout_editor_absoluteX="8dp" />
-
-    <com.stripe.android.view.CardInputWidget
-        android:id="@+id/card_input_widget"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        stripe:cardTint="@android:color/holo_purple"
-        stripe:cardTextErrorColor="@android:color/holo_orange_dark"
-        />
+        android:id="@+id/fragment"
+        android:name="com.mobewash.mobewash.PaymentFragment"
+        android:layout_width="373dp"
+        android:layout_height="509dp"
+        tools:layout_editor_absoluteX="7dp"
+        tools:layout_editor_absoluteY="-2dp"
+        tools:layout="@layout/fragment_payment" />
 
 </android.support.constraint.ConstraintLayout>
