+++ app/src/main/res/layout/fragment_payment.xml
@@ -1,14 +1,120 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:wallet="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    xmlns:stripe="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.mobewash.mobewash.PaymentFragment">
 
-    <!-- TODO: Update blank fragment layout -->
+<!--
 
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+    Android Pay Fragment
 
-</FrameLayout>
+    <fragment
+        android:id="@+id/wallet_fragment"
+        android:name="com.google.android.gms.wallet.fragment.SupportWalletFragment"
+        android:layout_width="335dp"
+        android:layout_height="39dp"
+        wallet:environment="test"
+        wallet:fragmentMode="buyButton"
+        tools:layout_editor_absoluteY="397dp"
+        tools:layout_editor_absoluteX="27dp"
+        tools:layout="@layout/fragment_payment" />
+
+    Card Input Widget
+    <com.stripe.android.view.CardInputWidget
+          android:id="@+id/card_input_widget"
+          android:layout_width="324dp"
+          android:layout_height="wrap_content"
+          stripe:cardTextErrorColor="@android:color/holo_red_dark"
+          stripe:cardTint="@android:color/holo_blue_light"
+          tools:layout_editor_absoluteX="74dp"
+          tools:layout_editor_absoluteY="201dp" />-->
+
+    <!--Car and Wash Details-->
+    <android.support.constraint.ConstraintLayout
+        android:layout_width="352dp"
+        android:layout_height="201dp"
+        tools:layout_editor_absoluteX="16dp"
+        tools:layout_editor_absoluteY="27dp">
+
+        <ImageView
+            android:id="@+id/CarView"
+            android:layout_width="162dp"
+            android:layout_height="165dp"
+            stripe:srcCompat="@drawable/common_google_signin_btn_icon_dark_normal"
+            tools:layout_editor_absoluteX="174dp"
+            tools:layout_editor_absoluteY="18dp" />
+
+        <ListView
+            android:id="@+id/DetailsView"
+            android:layout_width="125dp"
+            android:layout_height="170dp"
+            android:headerDividersEnabled="true"
+            tools:layout_editor_absoluteX="29dp"
+            tools:layout_editor_absoluteY="15dp" />
+    </android.support.constraint.ConstraintLayout>
+
+    <!--Book Button-->
+    <Button
+        android:id="@+id/BookButton"
+        android:layout_width="339dp"
+        android:layout_height="42dp"
+        android:text="Book"
+        tools:layout_editor_absoluteX="23dp"
+        tools:layout_editor_absoluteY="447dp" />
+
+    <!--Card Details-->
+    <LinearLayout
+        android:layout_width="352dp"
+        android:layout_height="120dp"
+        android:orientation="vertical"
+        tools:layout_editor_absoluteX="16dp"
+        tools:layout_editor_absoluteY="252dp">
+
+        <EditText
+            android:id="@+id/CardNumber"
+            android:layout_width="match_parent"
+            android:layout_height="40dp"
+            android:ems="10"
+            android:hint="Card Number"
+            android:inputType="numberDecimal"
+            android:singleLine="true"
+            tools:layout_editor_absoluteX="138dp"
+            tools:layout_editor_absoluteY="400dp" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="40dp"
+            android:orientation="horizontal">
+
+            <EditText
+                android:id="@+id/CVC"
+                android:layout_width="180dp"
+                android:layout_height="40dp"
+                android:ems="10"
+                android:hint="CVC"
+                android:inputType="numberDecimal"
+                tools:layout_editor_absoluteX="16dp"
+                tools:layout_editor_absoluteY="357dp" />
+
+            <EditText
+                android:id="@+id/ExpirationDate"
+                android:layout_width="180dp"
+                android:layout_height="40dp"
+                android:ems="10"
+                android:hint="MM/YY"
+                android:inputType="date" />
+        </LinearLayout>
+
+        <EditText
+            android:id="@+id/BillingAddress"
+            android:layout_width="match_parent"
+            android:layout_height="40dp"
+            android:ems="10"
+            android:hint="Billing Address"
+            android:inputType="textPostalAddress" />
+    </LinearLayout>
+
+</android.support.constraint.ConstraintLayout>
